Move View event handling logic into a dedicated file
